move ExchangeRates to core
