+++ core/Cargo.toml
@@ -18,12 +18,14 @@ crossbeam = "0.7"
 enum_primitive = "0.1"
 failure = "0.1"
 failure_derive = "0.1"
+itertools = "0.7"
 lazy_static = "1"
 lru-cache = "0.1"
 num = "0.2"
 num-bigint = "0.2"
 rand = "0.6"
 serde = "1"
+serde_json = "1"
 serde_derive = "1"
 siphasher = "0.2"
 uuid = { version = "0.6", features = ["serde", "v4"] }
