+++ core/src/core.rs
@@ -23,6 +23,7 @@ pub mod hash;
 pub mod id;
 pub mod merkle_proof;
 pub mod pmmr;
+pub mod price;
 pub mod transaction;
 pub mod verifier_cache;
 
@@ -35,6 +36,7 @@ pub use self::block_sums::*;
 pub use self::committed::Committed;
 pub use self::compact_block::*;
 pub use self::id::ShortId;
+pub use self::price::*;
 pub use self::transaction::*;
 
 /// Common errors
@@ -137,5 +139,4 @@ mod test {
 		assert!("5000000000.0" == amount_to_hr_string(5_000_000_000_000_000_000, true));
 		assert!("66.6" == amount_to_hr_string(66600000000, true));
 	}
-
 }
